Fix split CRLF handling
If the final character in the buffer was \r in
HTTPConnection::readLine(), it would leave it there to wait for the next
character to be added to the buffer. This is correct, but was unhandled
by the rest of the code. It would loop around in readLine() without ever
extending the buffer contents. The same in HTTPConnection::loop().
Result: an endless loop and a hung microcontroller.

The fix in HTTPConnection::loop() is to stop parsing further headers if
there is 0 or 1 further bytes available. Since even the empty line at
the end of the headers is 2 bytes (\r\n), this will never leave
available lines unread. If there is 1 byte in the buffer, the next
iteration of loop() will try to extend the buffer with new data.

This solution does not increase the maximum duration of a single
HTTPConnection::loop() iteration. We can't spin around reading further
network data, other things in the main loop might time out while we're
stuck there.